# techspecgen_swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tech Spec Generator API",
    "version": "1.0.0",
    "description": "Backend API for Tech Spec Generator - handles chat history storage with PostgreSQL and GCP integration",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    },
    {
      "url": "https://your-production-url.com",
      "description": "Production server"
    }
  ],
  "components": {
    "schemas": {
      "Message": {
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "model"
            ],
            "description": "The role of the message sender"
          },
          "content": {
            "type": "string",
            "description": "The content of the message"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Optional timestamp of the message"
          }
        }
      },
      "ChatSession": {
        "type": "object",
        "required": [
          "sessionId",
          "messages",
          "timestamp"
        ],
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "Unique identifier for the chat session"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "description": "Array of messages in the chat session"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the session was created"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "userAgent": {
                "type": "string"
              },
              "clientVersion": {
                "type": "string"
              }
            },
            "description": "Optional metadata about the session"
          }
        }
      },
      "SaveChatRequest": {
        "type": "object",
        "required": [
          "messages"
        ],
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "Optional session ID. If not provided, server will generate one."
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "minItems": 1,
            "description": "Array of messages (must not be empty)"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Optional timestamp. Defaults to current server time if not provided."
          }
        }
      },
      "SaveChatResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the operation was successful"
          },
          "sessionId": {
            "type": "string",
            "description": "The session ID (generated by server if not provided)"
          },
          "message": {
            "type": "string",
            "description": "Success message"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "Error type"
          },
          "message": {
            "type": "string",
            "description": "Detailed error message"
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "healthy",
              "degraded"
            ],
            "description": "Overall system health status"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "uptime": {
            "type": "number",
            "description": "Server uptime in seconds"
          },
          "database": {
            "type": "object",
            "properties": {
              "connected": {
                "type": "boolean",
                "description": "Database connection status"
              },
              "sessionCount": {
                "type": "number",
                "description": "Number of stored chat sessions"
              }
            }
          },
          "environment": {
            "type": "string",
            "description": "Current environment (development/production)"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Health",
      "description": "System health and monitoring endpoints"
    },
    {
      "name": "Chat Sessions",
      "description": "Chat session management endpoints"
    }
  ],
  "paths": {
    "/api/chat/save": {
      "post": {
        "summary": "Save a chat session",
        "description": "Save a new chat session or update an existing one. If sessionId is not provided, the server will generate one automatically.",
        "tags": [
          "Chat Sessions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveChatRequest"
              },
              "examples": {
                "withSessionId": {
                  "summary": "Save with custom session ID",
                  "value": {
                    "sessionId": "chat-custom-123",
                    "messages": [
                      {
                        "role": "user",
                        "content": "Hello"
                      },
                      {
                        "role": "model",
                        "content": "Hi there!"
                      }
                    ],
                    "timestamp": "2024-01-01T00:00:00.000Z"
                  }
                },
                "withoutSessionId": {
                  "summary": "Save without session ID (server generates)",
                  "value": {
                    "messages": [
                      {
                        "role": "user",
                        "content": "Hello"
                      },
                      {
                        "role": "model",
                        "content": "Hi there!"
                      }
                    ],
                    "timestamp": "2024-01-01T00:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Chat session saved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveChatResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (missing or empty messages array)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/{sessionId}": {
      "get": {
        "summary": "Get a specific chat session",
        "description": "Retrieve a chat session by its unique session ID",
        "tags": [
          "Chat Sessions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique session ID",
            "example": "chat-1234567890-abc123"
          }
        ],
        "responses": {
          "200": {
            "description": "Chat session retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ChatSession"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Chat session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat": {
      "get": {
        "summary": "Get all chat sessions",
        "description": "Retrieve all stored chat sessions, ordered by most recent first",
        "tags": [
          "Chat Sessions"
        ],
        "responses": {
          "200": {
            "description": "List of all chat sessions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "number",
                      "description": "Total number of sessions",
                      "example": 5
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ChatSession"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "summary": "Health check endpoint",
        "description": "Check the health status of the API and database connection",
        "tags": [
          "Health"
        ],
        "responses": {
          "200": {
            "description": "System is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          },
          "503": {
            "description": "System is degraded (database connection failed)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}